requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
blinker=1.9.0=pypi_0
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.11.26=hca03da5_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.4.0=pypi_0
click=8.1.7=pypi_0
flask=3.1.0=pypi_0
idna=3.10=pypi_0
itsdangerous=2.2.0=pypi_0
jinja2=3.1.4=pypi_0
libffi=3.4.4=hca03da5_1
markupsafe=3.0.2=pypi_0
ncurses=6.4=h313beb8_0
oauthlib=3.2.2=pypi_0
openssl=3.0.15=h80987f9_0
pip=24.2=py310hca03da5_0
python=3.10.16=hb885b13_1
python-dotenv=1.0.1=pypi_0
readline=8.2=h1a28f6b_0
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
setuptools=75.1.0=py310hca03da5_0
sqlite=3.45.3=h80987f9_0
tk=8.6.14=h6ba3021_0
tzdata=2024b=h04d1e81_0
urllib3=2.2.3=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.44.0=py310hca03da5_0
xz=5.4.6=h80987f9_1
zlib=1.2.13=h18a0788_1
